adds support for github codespaces for fast, portable dev env setup (plus minimal fixes) (#7)
Using GH Codespaces, a dev can launch a browser instance of VS Code running the full env needed for Semantic Workbench in just a few minutes, full automated. This PR also includes the necessary .vscode launch configuration for run/debug w/ the UI and a minimal/intro readme for this approach.

Also applied a few small fixes for issues encountered while testing up to running the canonical assistant, including the fixes submitted in the other open PR (at the time of opening this one).

Includes recommended settings and extensions that we've found useful in our dev of this solution, but since they are our opinions, can either remain simply as recommendations or removed before merging if that is preferred.