refactor inspectors to canvas components (#52)
Initial change to support to change the current "inspector" side panel
into a canvas that functions at different levels:
- Conversation canvas: contains a panel for all of the assistants within
a conversation, to host their individual canvas surfaces. It will (soon)
have a file canvas for managing/viewing/interacting with files at the
conversation level
- Assistant canvas: per-assistant canvas that has the inspectors by
default, but will start to extend with others that are more use-case
specific UX that works for specific data models being passed

The new components are coming separately; this is just the refactor to
the naming/location of the existing components for these new pieces.