Add pwsh equiv tool scripts (#87)
Added PowerShell equivalent tool scripts for resetting data, starting
the app, service, canonical, and the python2 sample chatbot.

I also changed the script for the chatbot sample to start on the same
port as the vscode launch.

I added a new script for PowerShell and bash to start the service,
chatbot, and app as a convenience for checking things are working
locally.

I also don't think we want scripts here for all of the examples. I
suggest removing the other ones and just keeping canonical and chatbot.
Example authors can choose to provide a script or not for starting from
the command line. There are enough examples here without more that
people can see how to do that or start other things from the cli if that
is how they prefer to work.