Simplifies assistant launch by picking random port
Now that assistants advertise their callback URL to the workbench, they
can use a random available port at startup.
Which means we don't have to pick a port for each assistant when we
start them, and make sure that none of their launch configurations
conflict.
This highlighted a bug in the workbench-service that wasn't re-creating
the client when the URL changed - this also fixes that bug