When using Aspire locally always use the same port, without proxying (#287)
Aspire by default randomizes ports, however this means that cookies from
a previous session most likely don't work, in which case the workbench
show the set of initial banners and asks to sign in, every time.

The PR changes this behavior on localhost, to always use port 4000
(assuming it's free), similarly to how the app runs without Aspire.