Skill library - Skill Assistant mechanics (#297)
This PR completes the base mechanics of the skill library and implements
it into the skill assistant. With this PR, the skill library now
supports the creation of skill packages, including actions and routines,
and allows for their registration to an assistant, as demonstrated in
the skill assistant.

To demo, create an instance of the skill assistant in the workbench and
send a "/help" message to the assistant.

Instruction routines and state machine routines are fully supported.