Filesystem MCP Server and sse query params (#351)
This adds a new filesystem MCP server written in Python modeled on the
one from Anthropic. This one has tests on the underlying operations so
we can make sure in CI it doesn't break. Anthropic does not test theirs,
particularly on Windows. I will work on a CI test after syncing up with
Mark and submit that later in a subsequent PR.

This also enables sending the MCP args in the config as a comma separated list in the query params as args=value,value for SSE connections. 

I have validated this on Windows and WSL. I validated the MCP server
tools by having an assistant configured with this read the test suite
then run similar operations from the chat and report the results.