Integrates memory-user-bio with codespace-assistant (#437)
- adds github workflow to auto-deploy mcp-server
- adds the mcp-server-memory-user-bio to the configured hosted MCP
servers for codespace-assistant
- adds new `prompts_to_auto_include` attribute to `MCPServerConfig`. for
an MCP server that has this configured, the prompt with the names in
this list will be automatically included in the system prompt. for this
mcp-server, the "user-bio" prompt is configured to be auto-included.
this prompt contains the memories.

also:
- cleans up a number of tests in other projects that either couldn't run
test discovery or tried to run on platforms where they will never work