Navigator hand off (#547)
When creating a conversation with an assistant, the navigator will send
a message providing context about the user's goal. Additionally, it will
copy relevant files to the new conversation.

The details about the message to send and which files to copy are
included in the metadata of the new conversation, and read and acted
upon in the on_conversation_created event handler.